+++ tests/functional/test_cli.py
@@ -521,7 +521,7 @@ class TestWeaverClient(TestWeaverClientBase):
         """
         result = self.run_execute_inputs_schema_variant("Execute_Echo_cwl_schema.yml", mock_exec=False)
         job_id = result.body["jobID"]
-        result = mocked_sub_requests(self.app, self.client.monitor, job_id, timeout=1, interval=1)
+        result = mocked_sub_requests(self.app, self.client.monitor, job_id, timeout=5, interval=1)
         assert result.success, result.text
         assert "undefined" not in result.message
         assert result.body.get("status") == Status.SUCCEEDED