+++ CHANGES.rst
@@ -12,6 +12,11 @@ Changes
 
 Changes:
 --------
+- Add more secure path validations steps before fetching contents.
+- Disallow ``builtin`` processes expecting a user-provided input path to run with local file references such that
+  they must respect any configured server-side remote file access rules instead of bypassing security validations
+  through resolved local paths.
+- Add multiple validation checks for more secure file paths handling when retrieving contents from remote locations.
 - Add more tests to validate core code paths of ``builtin`` `Process` ``jsonarray2netcdf``, ``metalink2netcdf`` and
   ``file_index_selector`` with validation of happy path and error handling conditions.